Update: Modularise Breadcrumb component
