+++ src/components/adslotUi/TreePicker/TreePickerNavComponent.jsx
@@ -1,5 +1,5 @@
 import React, { PropTypes } from 'react';
-import Breadcrumb from 'components/alexandria/BreadcrumbComponent';
+import Breadcrumb from 'components/alexandria/Breadcrumb/component';
 import Search from 'components/alexandria/SearchComponent';
 import SvgSymbol from 'components/alexandria/SvgSymbolComponent';
 import TreePickerPropTypes from 'helpers/propTypes/TreePickerPropTypes';
