+++ src/components/alexandria/Breadcrumb/styles.scss
@@ -1,6 +1,6 @@
-@import '../color';
-@import '../font-size';
-@import '../font-weight';
+@import 'styles/color';
+@import 'styles/font-size';
+@import 'styles/font-weight';
 
 .breadcrumb-component {
   $breadcrumb-height: 10px;
