+++ src/components/core.js
@@ -40,7 +40,7 @@ import fastStatelessWrapper from 'components/adslotUi/fastStatelessWrapper';
 import Alert from 'components/alexandria/Alert/component';
 import Avatar from 'components/alexandria/Avatar/component';
 import BorderedWell from 'components/alexandria/BorderedWell/component';
-import Breadcrumb from 'components/alexandria/BreadcrumbComponent';
+import Breadcrumb from 'components/alexandria/Breadcrumb/component';
 import Card from 'components/alexandria/CardComponent';
 import Empty from 'components/alexandria/EmptyComponent';
 import FlexibleSpacer from 'components/alexandria/FlexibleSpacerComponent';
